+++ packages/graph-plugins-shared/src/plugins/internals/plugin.ts
