+++ packages/graph-plugins/src/adjacency-lists/types.ts
@@ -7,7 +7,7 @@ import { CoreNode } from '@magic/graph-primitives/types';
 
 import { Ref } from 'vue';
 
-import { GraphPlugin } from '../../../graph-plugins-shared/dist/types/plugins/internals/types.ts';
+import { GraphPlugin } from '../../../graph-plugins-shared/dist/types/plugins/internals/plugin.ts';
 
 /**
  * a mapping of nodes to their neighbors where neighbors are the full node objects
