+++ packages/graph-plugins/src/history/types.ts
@@ -3,7 +3,7 @@ import { ComputedRef, Ref } from 'vue';
 import {
   GraphPlugin,
   WithLifecycle,
-} from '../../../graph-plugins-shared/dist/types/plugins/internals/types.ts';
+} from '../../../graph-plugins-shared/dist/types/plugins/internals/plugin.ts';
 import { HistoryEventMap } from './events.ts';
 
 type HistoryOption = {
