+++ packages/graph-plugins/src/node-label/types.ts
@@ -3,7 +3,7 @@ import { MaybeGetter } from '@magic/utils/maybeGetter/index';
 import {
   GraphPlugin,
   WithLifecycle,
-} from '../../../graph-plugins-shared/dist/types/plugins/internals/types.ts';
+} from '../../../graph-plugins-shared/dist/types/plugins/internals/plugin.ts';
 import { CanvasPlugin } from '../canvas/types.ts';
 import { FocusPlugin } from '../focus/types.ts';
